+++ io/ODiffIO.re
@@ -1,5 +1,4 @@
 module Bmp = Bmp;
-module Png_Bigarray = Png_Bigarray;
 module Png = Png;
 module Jpg = Jpg;
 module Tiff = Tiff;
