+++ io/dune
@@ -3,9 +3,4 @@
  (public_name odiff-io)
  (flags
   (-w -40 -w +26))
- (libraries
-  odiff-io.png
-  odiff-io.png_bigarray
-  odiff-io.jpg
-  odiff-io.bmp
-  odiff-io.tiff))
+ (libraries odiff-io.png odiff-io.jpg odiff-io.bmp odiff-io.tiff))
