+++ io/jpg/Jpg.re
@@ -4,7 +4,6 @@ type data = Array1.t(int32, int32_elt, c_layout);
 
 module IO = {
   type t = data;
-  type row = int;
   let buffer = ref(None);
 
   let loadImage = (filename): Odiff.ImageIO.img(t) => {
@@ -14,26 +13,12 @@ module IO = {
     {width, height, image};
   };
 
-  let readRow = (img: Odiff.ImageIO.img(t), y): row => y;
-
   let readDirectPixel = (~x, ~y, img: Odiff.ImageIO.img(t)) => {
     (img.image).{y * img.width + x};
   };
 
-  let readImgColor = (x, row: row, img: Odiff.ImageIO.img(t)) => {
-    readDirectPixel(~x, ~y=row, img);
-  };
-
-  let setImgColor = (x, y, (r, g, b), img: Odiff.ImageIO.img(t)) => {
-    let a = (255 land 0xFF) lsl 24;
-    let b = (b land 0xFF) lsl 16;
-    let g = (g land 0xFF) lsl 8;
-    let r = (r land 0xFF) lsl 0;
-    Array1.set(
-      img.image,
-      y * img.width + x,
-      Int32.of_int(a lor b lor g lor r),
-    );
+  let setImgColor = (~x, ~y, color, img: Odiff.ImageIO.img(t)) => {
+    Array1.unsafe_set(img.image, y * img.width + x, color);
   };
 
   let saveImage = (img: Odiff.ImageIO.img(t), filename) => {
