+++ io/png/ReadPng.c
@@ -1,191 +1,112 @@
 #define CAML_NAME_SPACE
-#include <stdio.h>
-#include <string.h>
-#include <png.h>
+
+#include <spng.h>
+
 #include <caml/mlvalues.h>
 #include <caml/alloc.h>
 #include <caml/memory.h>
 #include <caml/fail.h>
 #include <caml/bigarray.h>
 
-char *concat(const char *s1, const char *s2)
-{
-  const size_t len1 = strlen(s1);
-  const size_t len2 = strlen(s2);
-  char *result = malloc(len1 + len2 + 1); // +1 for the null-terminator
-
-  memcpy(result, s1, len1);
-  memcpy(result + len1, s2, len2 + 1); // +1 to copy the null-terminator
-  return result;
-}
-
 CAMLprim value
-read_png_file_to_tuple(value file)
+read_png_file(value file)
 {
   CAMLparam1(file);
   CAMLlocal2(res, ba);
 
-  int width, height;
-  png_byte color_type;
-  png_byte bit_depth;
-  png_bytep *row_pointers = NULL;
-
+  int result = 0;
+  FILE *png;
+  spng_ctx *ctx = NULL;
+  unsigned char *out = NULL;
   const char *filename = String_val(file);
-  FILE *fp = fopen(filename, "rb");
 
-  png_structp png = png_create_read_struct(PNG_LIBPNG_VER_STRING, NULL, NULL, NULL);
-  if (!png)
-    caml_failwith(concat("Can not read the file -", filename));
-
-  png_infop info = png_create_info_struct(png);
-  if (!info)
-    caml_failwith(concat("Incorrect file type -", filename));
-
-  if (setjmp(png_jmpbuf(png)))
-    caml_failwith(concat("Can not read the file -", filename));
-
-  png_init_io(png, fp);
-
-  png_read_info(png, info);
-
-  width = png_get_image_width(png, info);
-  height = png_get_image_height(png, info);
-  color_type = png_get_color_type(png, info);
-  bit_depth = png_get_bit_depth(png, info);
-
-  // Read any color_type into 8bit depth, RGBA format.
-  // See http://www.libpng.org/pub/png/libpng-manual.txt
-
-  if (bit_depth == 16)
-    png_set_strip_16(png);
-
-  if (color_type == PNG_COLOR_TYPE_PALETTE)
-    png_set_palette_to_rgb(png);
-
-  // PNG_COLOR_TYPE_GRAY_ALPHA is always 8 or 16bit depth.
-  if (color_type == PNG_COLOR_TYPE_GRAY && bit_depth < 8)
-    png_set_expand_gray_1_2_4_to_8(png);
-
-  if (png_get_valid(png, info, PNG_INFO_tRNS))
-    png_set_tRNS_to_alpha(png);
-
-  // These color_type don't have an alpha channel then fill it with 0xff.
-  if (color_type == PNG_COLOR_TYPE_RGB ||
-      color_type == PNG_COLOR_TYPE_GRAY ||
-      color_type == PNG_COLOR_TYPE_PALETTE)
-    png_set_filler(png, 0xFF, PNG_FILLER_AFTER);
-
-  if (color_type == PNG_COLOR_TYPE_GRAY ||
-      color_type == PNG_COLOR_TYPE_GRAY_ALPHA)
-    png_set_gray_to_rgb(png);
-
-  png_set_interlace_handling(png);
-
-  png_read_update_info(png, info);
-
-  row_pointers = (png_bytep *)malloc(sizeof(png_bytep) * height);
-  int rowbytes = png_get_rowbytes(png, info);
-
-  for (int y = 0; y < height; y++)
-  {
-    row_pointers[y] = (png_byte *)malloc(rowbytes);
+  png = fopen(filename, "rb");
+  if(png == NULL) {
+    caml_failwith("error opening input file");
   }
 
-  png_read_image(png, row_pointers);
+  ctx = spng_ctx_new(0);
 
-  fclose(fp);
+  if(ctx == NULL) {
+    caml_failwith("spng_ctx_new() failed");
+    spng_ctx_free(ctx);
+    free(out);
+  }
 
-  png_destroy_read_struct(&png, &info, NULL);
+  /* Ignore and don't calculate chunk CRC's */
+  spng_set_crc_action(ctx, SPNG_CRC_USE, SPNG_CRC_USE);
+
+  /* Set memory usage limits for storing standard and unknown chunks,
+      this is important when reading untrusted files! */
+  size_t limit = 1024 * 1024 * 64;
+  spng_set_chunk_limits(ctx, limit, limit);
+
+  /* Set source PNG */
+  spng_set_png_file(ctx, png);
+
+  struct spng_ihdr ihdr;
+  result = spng_get_ihdr(ctx, &ihdr);
+
+  if(result) {
+    caml_failwith("spng_get_ihdr() error!");
+    spng_ctx_free(ctx);
+    free(out);
+  }
+
+  size_t out_size, out_width;
+  result = spng_decoded_image_size(ctx, SPNG_FMT_RGBA8, &out_size);
+  if(result) {
+    spng_ctx_free(ctx);
+  };
+
+  /* This is required to initialize for progressive decoding */
+  result = spng_decode_image(ctx, NULL, 0, SPNG_FMT_RGBA8, SPNG_DECODE_PROGRESSIVE);
+  if(result) {
+    spng_ctx_free(ctx);
+    caml_failwith(spng_strerror(result));
+  }
+
+  /* ihdr.height will always be non-zero if spng_get_ihdr() succeeds */
+  out_width = out_size / ihdr.height;
+
+  out = malloc(out_size);
+  if(out == NULL) {
+    spng_ctx_free(ctx);
+    free(out);
+  };
+  struct spng_row_info row_info = {0};
+
+  do {
+    result = spng_get_row_info(ctx, &row_info);
+    if(result) break;
+
+    result = spng_decode_row(ctx, out + row_info.row_num * out_width, out_width);
+  }
+  while(!result);
+
+  if(result != SPNG_EOI) {
+    caml_failwith(spng_strerror(result));
+    spng_ctx_free(ctx);
+    free(out);
+  }
 
   res = caml_alloc(4, 0);
+  ba = caml_ba_alloc_dims(CAML_BA_INT32 | CAML_BA_C_LAYOUT, 1, out, out_size);
 
-  Store_field(res, 0, Val_int(width));
-  Store_field(res, 1, Val_int(height));
-  Store_field(res, 2, Val_int(rowbytes));
-  Store_field(res, 3, Val_bp(row_pointers));
+  Store_field(res, 0, Val_int(ihdr.width));
+  Store_field(res, 1, Val_int(ihdr.height));
+  Store_field(res, 2, ba);
+  Store_field(res, 3, Val_bp(out));
+
+  spng_ctx_free(ctx);
 
   CAMLreturn(res);
 }
 
 CAMLprim value
-row_pointers_to_bigarray(png_bytep *row_pointers, value rowbytes_val, value height_val, value width_val)
+cleanup_png(value buffer)
 {
-  CAMLparam3(rowbytes_val, height_val, width_val);
-
-  int width = Int_val(width_val);
-  int height = Int_val(height_val);
-  int rowbytes = Int_val(rowbytes_val);
-
-  unsigned char *total_pixels = malloc(height * rowbytes);
-
-  for (int y = 0; y < height; y++)
-  {
-    memcpy(total_pixels + y * rowbytes, row_pointers[y], rowbytes);
-  }
-
-  CAMLreturn(caml_ba_alloc_dims(CAML_BA_INT32 | CAML_BA_C_LAYOUT, 1, total_pixels, width * height));
-}
-
-CAMLprim value
-create_empty_img(value height_val, value width_val)
-{
-  CAMLparam2(height_val, width_val);
-  int width = Int_val(width_val);
-  int height = Int_val(height_val);
-
-  png_bytep *row_pointers = (png_bytep *)malloc(sizeof(png_bytep) * height);
-
-  for (int y = 0; y < height; y++)
-  {
-    row_pointers[y] = (png_byte *)malloc(width * 4); // we always use RGBA
-  }
-
-  CAMLreturn(Val_bp(row_pointers));
-}
-
-CAMLprim value
-read_row(png_bytep *row_pointers, value y_val, value img_width_val)
-{
-  CAMLparam2(y_val, img_width_val);
-  int y = Int_val(y_val);
-  int img_width = Int_val(img_width_val);
-
-  png_bytep row = row_pointers[y];
-
-  CAMLreturn(caml_ba_alloc_dims(CAML_BA_INT32 | CAML_BA_C_LAYOUT, 1, row, img_width));
-}
-
-CAMLprim value
-set_pixel_data(png_bytep *row_pointers, value x_val, value y_val, value pixel_val)
-{
-  CAMLparam3(x_val, y_val, pixel_val);
-
-  int x = Int_val(x_val);
-  int y = Int_val(y_val);
-
-  png_bytep row = row_pointers[y];
-  png_bytep px = &(row[x * 4]);
-
-  px[0] = Int_val(Field(pixel_val, 0));
-  px[1] = Int_val(Field(pixel_val, 1));
-  px[2] = Int_val(Field(pixel_val, 2));
-  px[3] = 255;
-
-  CAMLreturn(Val_unit);
-}
-
-CAMLprim value
-free_row_pointers(png_bytep *row_pointers, value height_value)
-{
-  CAMLparam1(height_value);
-  int height = Int_val(height_value);
-
-  for (int y = 0; y < height; y++)
-  {
-    free(row_pointers[y]);
-  }
-
-  free(row_pointers);
-
+  CAMLparam1(buffer);
+  free(Bp_val(buffer));
   CAMLreturn(Val_unit);
 }
