--- io/png_bigarray/ReadPng_Bigarray.c
@@ -1,112 +1,0 @@
-#define CAML_NAME_SPACE
-
-#include <spng.h>
-
-#include <caml/mlvalues.h>
-#include <caml/alloc.h>
-#include <caml/memory.h>
-#include <caml/fail.h>
-#include <caml/bigarray.h>
-
-CAMLprim value
-read_spng_file(value file)
-{
-  CAMLparam1(file);
-  CAMLlocal2(res, ba);
-
-  int result = 0;
-  FILE *png;
-  spng_ctx *ctx = NULL;
-  unsigned char *out = NULL;
-  const char *filename = String_val(file);
-
-  png = fopen(filename, "rb");
-  if(png == NULL) {
-    caml_failwith("error opening input file");
-  }
-
-  ctx = spng_ctx_new(0);
-
-  if(ctx == NULL) {
-    caml_failwith("spng_ctx_new() failed");
-    spng_ctx_free(ctx);
-    free(out);
-  }
-
-  /* Ignore and don't calculate chunk CRC's */
-  spng_set_crc_action(ctx, SPNG_CRC_USE, SPNG_CRC_USE);
-
-  /* Set memory usage limits for storing standard and unknown chunks,
-      this is important when reading untrusted files! */
-  size_t limit = 1024 * 1024 * 64;
-  spng_set_chunk_limits(ctx, limit, limit);
-
-  /* Set source PNG */
-  spng_set_png_file(ctx, png);
-
-  struct spng_ihdr ihdr;
-  result = spng_get_ihdr(ctx, &ihdr);
-
-  if(result) {
-    caml_failwith("spng_get_ihdr() error!");
-    spng_ctx_free(ctx);
-    free(out);
-  }
-
-  size_t out_size, out_width;
-  result = spng_decoded_image_size(ctx, SPNG_FMT_RGBA8, &out_size);
-  if(result) {
-    spng_ctx_free(ctx);
-  };
-
-  /* This is required to initialize for progressive decoding */
-  result = spng_decode_image(ctx, NULL, 0, SPNG_FMT_RGBA8, SPNG_DECODE_PROGRESSIVE);
-  if(result) {
-    spng_ctx_free(ctx);
-    caml_failwith(spng_strerror(result));
-  }
-
-  /* ihdr.height will always be non-zero if spng_get_ihdr() succeeds */
-  out_width = out_size / ihdr.height;
-
-  out = malloc(out_size);
-  if(out == NULL) {
-    spng_ctx_free(ctx);
-    free(out);
-  };
-  struct spng_row_info row_info = {0};
-
-  do {
-    result = spng_get_row_info(ctx, &row_info);
-    if(result) break;
-
-    result = spng_decode_row(ctx, out + row_info.row_num * out_width, out_width);
-  }
-  while(!result);
-
-  if(result != SPNG_EOI) {
-    caml_failwith(spng_strerror(result));
-    spng_ctx_free(ctx);
-    free(out);
-  }
-
-  res = caml_alloc(4, 0);
-  ba = caml_ba_alloc_dims(CAML_BA_INT32 | CAML_BA_C_LAYOUT, 1, out, out_size);
-
-  Store_field(res, 0, Val_int(ihdr.width));
-  Store_field(res, 1, Val_int(ihdr.height));
-  Store_field(res, 2, ba);
-  Store_field(res, 3, Val_bp(out));
-
-  spng_ctx_free(ctx);
-
-  CAMLreturn(res);
-}
-
-CAMLprim value
-cleanup_png_bigarray(value buffer)
-{
-  CAMLparam1(buffer);
-  free(Bp_val(buffer));
-  CAMLreturn(Val_unit);
-}
